fix: decimal numbers on wallets

+++ src/data/entities/wallet.ts
@@ -11,7 +11,7 @@ export class WalletModel {
   @Column({ name: 'address', type: 'varchar', nullable: false })
   address: string;
 
-  @Column({ name: 'lastBalance', type: 'int', nullable: true, default: 0 })
+  @Column({ name: 'lastBalance', type: 'float', nullable: true, default: 0 })
   lastBalance?: Number;
 
   // @Column({ name: 'name', type: 'varchar', nullable: false })
